Revert "Deduplicate overlay file upload handler"
