+++ src/components/canvas/ModelDetailsDialog.tsx
@@ -85,7 +85,6 @@ interface ModelDetailsDialogProps {
   onOpenChange: (open: boolean) => void;
   model: MediaModel | null;
   onSave?: (parameters: Record<string, any>) => void;
-  initialParameters?: Record<string, any>;
 }
 
 export function ModelDetailsDialog({
@@ -93,7 +92,6 @@ export function ModelDetailsDialog({
   onOpenChange,
   model,
   onSave,
-  initialParameters,
 }: ModelDetailsDialogProps) {
   const [parameters, setParameters] = useState<Record<string, any>>({});
 
@@ -101,16 +99,11 @@ export function ModelDetailsDialog({
     if (model?.parameters) {
       const initialParams: Record<string, any> = {};
       model.parameters.forEach((param) => {
-        const existingValue = initialParameters?.[param.name];
-        if (existingValue !== undefined) {
-          initialParams[param.name] = existingValue;
-        } else {
-          initialParams[param.name] = param.default || "";
-        }
+        initialParams[param.name] = param.default || "";
       });
       setParameters(initialParams);
     }
-  }, [initialParameters, model]);
+  }, [model]);
 
   if (!model) {
     return null;
